♻️ Rename directory sqlslogopts to opts
but package name is still sqlslogopts
