+++ internal/wrap/open.go
@@ -6,7 +6,7 @@ import (
 	"database/sql/driver"
 	"log/slog"
 
-	"github.com/akm/sql-slog/sqlslogopts"
+	sqlslogopts "github.com/akm/sql-slog/opts"
 )
 
 /*
