+++ opts/conn.go
