+++ opts/conn_test.go
