+++ opts/connector.go
