+++ opts/connector_test.go
