+++ opts/id_gen_chacha8.go
