+++ opts/id_gen_chacha8_test.go
