+++ opts/logger_test.go
