+++ opts/rows.go
