+++ opts/rows_test.go
