+++ opts/slog.go
