+++ opts/slog_test.go
